refactor(models): move CatalogPlaylist from core:catalog to core:models/commonMain
