+++ core/catalog/src/main/kotlin/com/riffle/core/catalog/CatalogPlaylist.kt
@@ -1,10 +1,4 @@
 package com.riffle.core.catalog
 
-data class CatalogPlaylist(
-    val id: String,
-    val rootId: String,
-    val name: String,
-    val bookCount: Int,
-    /** Ordered item ids in the playlist; empty when the caller asked for summary only. */
-    val itemIds: List<String> = emptyList(),
-)
+@Deprecated("Moved to core:models", ReplaceWith("com.riffle.core.models.CatalogPlaylist"))
+typealias CatalogPlaylist = com.riffle.core.models.CatalogPlaylist
