+++ core/models/src/commonMain/kotlin/com/riffle/core/models/CatalogPlaylist.kt
@@ -1,0 +1,10 @@
+package com.riffle.core.models
+
+data class CatalogPlaylist(
+    val id: String,
+    val rootId: String,
+    val name: String,
+    val bookCount: Int,
+    /** Ordered item ids in the playlist; empty when the caller asked for summary only. */
+    val itemIds: List<String> = emptyList(),
+)
